Ignore prefix key inside pasted text
It prevents interpretation of a character that is set as a prefix key
when pasting text from outside of tmux (mouse middle-click,
ctrl-shift-v, etc...).

It allows setting any printable character as a prefix and still being
able to paste text including this character without a risk of messing up
tmux window/session.

For example now you can use backtick character (`) as a prefix and you
can safely paste bash code that uses backticks.